requirements.txt
Pillow>=10.0.0
pyperclip>=1.8.2
customtkinter>=5.2.0